+++ AGENTS.md
@@ -34,6 +34,7 @@ Tests use Vitest, Happy DOM, Testing Library, and WXT's Vitest plugin. Name file
 - Keep snapshot raw-card access and note traversal in `infrastructure/storage/snapshot.ts`. Share raw sync configuration/status and data timestamp access through `infrastructure/storage/sync-metadata.ts` across sync, backup/reset, and data tracking. Reuse stats and note APIs where they preserve raw fields; keep cross-entity ordering, PAT preservation, and imported timestamp restoration in services. Snapshot helpers do not mark local edits.
 - Keep backup payload contracts and JSON serialization in `infrastructure/storage/backup-codec.ts`; keep structure/schema acceptance, shallow validation, and legacy settings conversion in `domain/backup-import.ts`. Domain policy passes record contents through without importing storage types. Import/export services retain schema reads, conditional clock reads, and reset/restore orchestration.
 - Keep Octokit construction and requests in `infrastructure/github/client.ts`. Sync services choose client lifetime and retain validation, response/error handling, localization reads, timestamps, module-level sync state, and network execution inside the write queue.
+- Keep content mounting and orchestration in `content/bootstrap.ts`; `entrypoints/content.ts` owns WXT registration and startup. Preserve ping, translation loading, mounting/observer startup, auto-reset order, and ignored disposers; lifecycle fixes remain separate.
 - Route schema changes through a new, sequential migration in `infrastructure/storage/migrations.ts`.
 - Use `formatLocalDate` and `isDueByDate` from `domain/review-day.ts` with explicit dates and `dayStartHour` for review-day comparisons; do not compare raw timestamps.
 - Keep clock/settings reads and persistence sequencing in services; domain calculations take explicit inputs. Preserve the service-owned FSRS instance and parameters.
